return int

+++ app/Services/PepecoinExplorerService.php
@@ -152,7 +152,7 @@ class PepecoinExplorerService
             $this->getCacheKey(__FUNCTION__),
             Carbon::now()->addHours(4),
             function (): int {
-                return (new Collection($this->rpcService->getBlockchainInfo()))
+                 return (int) (new Collection($this->rpcService->getBlockchainInfo()))
                     ->get('size_on_disk', 0);
             }
         );
